refactor: replace ReverseSerializedPageReader with `with_reverse_pages` flag
Per @alamb's review (#9937 review-4255127096): the standalone
`ReverseSerializedPageReader` was ~95% duplicate of
`SerializedPageReader::Pages` mode — the only structural difference was
`pop_back` vs `pop_front`. Maintaining two near-identical readers adds
ongoing complexity without any benefit visible to downstream callers.

This collapses the two paths into one:

  - `SerializedPageReaderState::Pages` gains a `reverse: bool` field
  - new builder method `SerializedPageReader::with_reverse_pages(true)`
    flips the flag (and seeds `page_index` to `len - 1` so the page
    ordinal stays correct, including for encryption AAD)
  - `get_next_page` / `peek_next_page` / `skip_next_page` /
    `peek_next_page_offset` switch `pop_front`/`front` to
    `pop_back`/`back` and `+= 1` to `saturating_sub(1)` under the flag
  - the dictionary page is still emitted first in both modes (data
    pages depend on the loaded dictionary)

The dedicated `reverse_serialized_reader.rs` module is removed (~1080
lines deleted). Net change: -190 lines source, single code path.

The 21 reverse-correctness tests are migrated to a `reverse_pages_tests`
sub-module in `serialized_reader.rs` and updated to use the new builder
API. One additional test is added (`with_reverse_pages_is_noop_in_values_mode`)
verifying the flag is harmless when the reader is constructed in
`Values` mode (no `OffsetIndex`).

The `reverse_page_reader` criterion bench is updated to use the new
API. Bench numbers are within run-to-run noise of the pre-refactor
result: ~50x speedup for `n <= one page`, scaling down with `n` as
expected.

Tracking: <https://github.com/apache/arrow-rs/issues/9934>
